pyproject.toml
[project]
name = "aiops-agent"
version = "0.1.0"
description = "Knowledge-base + AIOps multi-agent system (Python + RAG + LangGraph + Kafka + Neo4j + Milvus + Redis Stack)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "loguru>=0.7",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "tenacity>=8.2",
    "numpy>=1.26",
    "confluent-kafka>=2.5",
    "redis>=5.0",
    "neo4j>=5.20",
    "pymilvus>=2.4",
    "rank-bm25>=0.2.2",
    "jieba>=0.42",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.2",
    "sentence-transformers>=3.0",
    "FlagEmbedding>=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "testcontainers>=4.5",
]
eval = [
    "ragas>=0.2",
    "datasets>=2.20",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["eval: marks tests as evaluation (deselect with '-m \"not eval\"')"]

[tool.ruff]
line-length = 110
target-version = "py311"
